single file configuration
Configuration no longer has to be split in 4 parts. Users can specify as
many or as little files as they want, provided that the content of those
files still define all 4 components.

In order to prevent selecting two components of the same time,
components now has to be explicitly created and named.

Network component has been renamed to Laboratory, and network hosts to
devices.