variables substition in commands
Before sending commands all instances in form @{host.<attr>} and
@{host.<interface>.<attr>} will be replaced with their value in
referenced model host or the bound device (or interface) if not present
in the model.

The 'name' attribute is treated differently as we don't want names from
model to be present in the commands sent so we're always using the bound
object's name instead.